backend/requirements.txt
Flask
Flask-Cors
psycopg2-binary
scikit-learn
tensorflow
pandas
numpy
geopandas
matplotlib
gunicorn
python-dotenv
requests